LayoutTests/media/media-source/media-source-play-expected.txt


EXPECTED (source.readyState == 'closed') OK
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
RUN(sourceBuffer.appendBuffer(sample))
EVENT(updateend)
RUN(source.duration = 1)
RUN(source.endOfStream())
RUN(video.play())
EVENT(ended)
END OF TEST
